Add redis_url from external app
