+++ authentication/config.py
@@ -1,0 +1,5 @@
+"""
+    authentication.config
+    ~~~~~~~~~~~~
+    Implements the configuration related objects.
+"""
